HtmlBridge Phase 3: split AnchorFunctions.cs under the 750-line ratchet
The Phase-3 line-limit guard (No_New_HtmlBridge_Production_File_Exceeds_The_Line_Limit)
caught its first new over-limit file: AnchorResolver/AnchorFunctions.cs grew 748 -> 767
lines as the Phase 5 native anchor() / @position-try handoff work added the MVP-subset
gating predicates, tripping the ratchet.

Rather than add an exemption, move the cohesive native-handoff cluster
(IsMvpNativeAnchorInsetBox / ...SizeBox / ...CombinedBox, IsCombinedAnchorRefNative,
NativePositionTryHandoffSupported / AxisSizeHandoffSupported and the HasInset /
IsAutoLength / IsEngineSizedIntrinsic helpers) into a sibling DomBridge partial
AnchorFunctions.NativeHandoff.cs (455 lines). AnchorFunctions.cs drops to 326 lines,
keeping the ResolveAnchorFunctions walk, intervening-scroll and anchor-size() resolution,
and the small predicate helpers.

Pure partial-class relocation: no signature, accessibility, or logic change, so
behaviour-identical by construction. Broiler.HtmlBridge.Dom builds clean and the guard
is green (0 offenders / 0 stale exemptions). The exemption list is unchanged.
